[TASK] move main adapt method to abstract, since it more general for usage. Added transformSourceData method to manipulate data before adapting
